Prepare deployment for beta releases

+++ .github/deploy.sh
@@ -11,10 +11,15 @@ cp util/gh-pages/index.html out/master
 python3 ./util/export.py out/master/lints.json
 
 if [[ -n $TAG_NAME ]]; then
+  if [[ $TAG_NAME == beta-* ]]; then
+    TAG_NAME="beta"
+  fi
   echo "Save the doc for the current tag ($TAG_NAME) and point stable/ to it"
-  cp -r out/master "out/$TAG_NAME"
-  rm -f out/stable
-  ln -s "$TAG_NAME" out/stable
+  cp -Tr out/master "out/$TAG_NAME"
+  if [[ $TAG_NAME != "beta" ]]; then
+    rm -f out/stable
+    ln -s "$TAG_NAME" out/stable
+  fi
 fi
 
 # Generate version index that is shown as root index page
@@ -35,7 +40,7 @@ fi
 
 if [[ -n $TAG_NAME ]]; then
   # Add the new dir
-  git add $TAG_NAME
+  git add "$TAG_NAME"
   # Update the symlink
   git add stable
   # Update versions file
